Create a Ruby Hash data structure
Ruby does data structures pretty well, arrays handle linked lists, queues, and stacks with built in methods, it's up to the programmer to stick to the behavior of those data structures.  Yes, I know a built in Hash class exists for Ruby, but I wanted as an exercise to build my own version to get a feel for the language and practice some data structures.